requirements.txt
Flask
Flask-SQLAlchemy
pandas
scikit-learn
joblib
requests
python-dotenv
apscheduler
folium
twilio
tensorflow
